TRegex: documentation

+++ dataset/GitHub_Java/oracle.graal/regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/automaton/SimpleStateIndex.java
@@ -67,14 +67,14 @@ public abstract class SimpleStateIndex<T> implements StateIndex<T>, Iterable<T> 
 
     public void add(T state) {
         assert !states.contains(state);
-        setStateId(state, states.size());
+        setStateId(state, (short) states.size());
         states.add(state);
         assert states.get(getStateId(state)) == state;
     }
 
-    protected abstract int getStateId(T state);
+    protected abstract short getStateId(T state);
 
-    protected abstract void setStateId(T state, int id);
+    protected abstract void setStateId(T state, short id);
 
     @Override
     public int getNumberOfStates() {
@@ -82,8 +82,8 @@ public abstract class SimpleStateIndex<T> implements StateIndex<T>, Iterable<T> 
     }
 
     @Override
-    public int getId(T state) {
-        int id = getStateId(state);
+    public short getId(T state) {
+        short id = getStateId(state);
         assert states.get(id) == state;
         return id;
     }
