+++ dataset/GitHub_Java/oracle.graal/regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/automaton/SmallStateSetImpl.java
@@ -90,7 +90,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         return toBit(stateIndex.getId((S) o));
     }
 
-    private static long toBit(int id) {
+    private static long toBit(short id) {
         assert 0 <= id && id < 64;
         return 1L << id;
     }
@@ -108,7 +108,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         return notPresent;
     }
 
-    private void removeId(int id) {
+    private void removeId(short id) {
         assert (set & toBit(id)) != 0;
         set &= ~toBit(id);
     }
@@ -260,7 +260,7 @@ final class SmallStateSetImpl<S> implements StateSet<S> {
         @Override
         public void remove() {
             assert bitIndex > 0;
-            stateSet.removeId(bitIndex - 1);
+            stateSet.removeId((short) (bitIndex - 1));
         }
     }
 }
